requirements.txt
streamlit==0.89.0
pillow==8.3.1
pytesseract==0.3.8
pandas==1.3.3